refactor: straight path
