+++ library/lib/constants/edgeConstants.ts
@@ -6,3 +6,4 @@ export const MARKER_PADDING = -3 // Reduce padding to bring marker closer to the
 export const SOURCE_CONNECTION_POINT_PADDING = 3.5 // Reduce padding to bring marker closer to the element
 export const STEP_BOARDER_RADIUS = 0 //Radius of the corners of Smooth Edge
 export const USECASE_PADDING = 5
+export const EDGE_HIGHTLIGHT_STROKE_WIDTH = 15
