+++ library/lib/edges/edgeTypes/ClassDiagramEdge.tsx
@@ -12,6 +12,7 @@ import { useDiagramStore, usePopoverStore } from "@/store/context"
 import { useShallow } from "zustand/shallow"
 import { useToolbar } from "@/hooks"
 import { useRef } from "react"
+import { EDGE_HIGHTLIGHT_STROKE_WIDTH } from "@/constants"
 
 export const ClassDiagramEdge = ({
   id,
@@ -111,7 +112,7 @@ export const ClassDiagramEdge = ({
           className="edge-overlay"
           d={overlayPath}
           fill="none"
-          strokeWidth={12}
+          strokeWidth={EDGE_HIGHTLIGHT_STROKE_WIDTH}
           pointerEvents="stroke"
           style={{
             opacity: isReconnectingRef.current ? 0 : 0.4,
