+++ library/lib/hooks/index.ts
@@ -11,3 +11,4 @@ export * from "./useKeyboardShortcuts"
 export * from "./useDragOver"
 export * from "./useHandleFinder"
 export * from "./useStepPathEdge"
+export * from "./useStraightPathEdge"
